+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Release History
 
 ## v9.0.2
+> Released: 2026-04-xx
+
 - [Changed] Handle multiple ip address retrieval during New-ExtractDataFromSDDCBackup
 
 ## v9.0.1
-
 > Released: 2025-09-23
 
-- Initial release of `VMware.CloudFoundation.InstanceRecovery` for VCF 9.x.
+- [Added] Initial release of `VMware.CloudFoundation.InstanceRecovery` for VCF 9.x.