+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v9.0.2
 > Released: 2026-04-xx
 
+- [Added] Timings to Functions
 - [Changed] Handle multiple ip address retrieval during New-ExtractDataFromSDDCBackup
 
 ## v9.0.1