+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - [Added] New-VVFBasedPartialBringupValidation and New-VVFBasedPartialBringup functions
 - [Changed] Handle multiple ip address retrieval during New-ExtractDataFromSDDCBackup
 - [Changed] IP Pools for VVF Partial bringup now handle multiple ip ranges instead of just one
+- [Changed] DNS Resolution failures handled gracefully and warning given
 
 ## v9.0.1
 > Released: 2025-09-23